refactor: split pydantic schemas into modules
